requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.40
asyncpg==0.30.0
pydantic==2.11.3
pydantic-settings==2.8.1
email-validator==2.2.0
python-jose[cryptography]==3.4.0
bcrypt==4.3.0
python-multipart==0.0.20
httpx==0.28.1
